requirements.txt
gunicorn==19.9.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
bcrypt==3.1.4
cffi==1.11.5
click==6.7
colorama==0.3.9
coverage==4.5.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.6
Flask-Testing==0.7.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
pathlib2==2.3.2
pep8==1.7.1
pluggy==0.7.1
psycopg2==2.7.5
py==1.6.0
pycparser==2.19
PyJWT==1.6.4
pylint==2.1.1
pytest==3.8.0
pytest-cov==2.6.0
scandir==1.9.0
six==1.11.0
typed-ast==1.1.0
typing==3.6.6
Werkzeug==0.14.1
wrapt==1.10.11